+++ packages/at_secondary_server/lib/src/connection/outbound/outbound_message_listener.dart
@@ -19,7 +19,8 @@ class OutboundMessageListener {
   /// Listens to the underlying connection's socket if the connection is created.
   /// @throws [AtConnectException] if the connection is not yet created
   void listen() async {
-    logger.finest('Calling outbound underlying.listen within runZonedGuarded block');
+    logger.finest(
+        'Calling outbound underlying.listen within runZonedGuarded block');
 
     runZonedGuarded(() {
       outboundClient.outboundConnection?.underlying.listen(_messageHandler,
@@ -131,6 +132,7 @@ class OutboundMessageListener {
 
   /// Closes the [OutboundClient]
   void _finishedHandler() async {
+    logger.info('_finishedHandler called - closing connection');
     _closeOutboundClient();
   }